Introduce unit tests (first testing enforcer API with OPA query mocks)
